requirements.txt
fastapi
uvicorn
Pillow
python-multipart
torch
torchvision
numpy